[M] 合并上游代码至 d896462c2db40126938a2690c47827df7a177028

+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/abstractItems/AReactor.java
@@ -318,7 +318,7 @@ public abstract class AReactor extends AbstractEnergyGenerator {
 
                         if (space >= produced) {
                             ChargableBlock.addCharge(l, getEnergyProduction());
-                            return charge + getEnergyProduction();
+                            return (double) (charge + getEnergyProduction());
                         } else {
                             return charge;
                         }
